move all myVidoeCourses page components to components folder
